+++ packages/serron/csrc/bindings.cpp
@@ -7,6 +7,8 @@
 TORCH_LIBRARY_EXPAND(TORCH_EXTENSION_NAME, ops) {
     ops.def("erode(Tensor input, Tensor kernel, int border) -> Tensor");
     ops.def("dilate(Tensor input, Tensor kernel, int border) -> Tensor");
+    ops.def("erode_backward(Tensor grad_output, Tensor input, Tensor kernel, int border) -> (Tensor, Tensor)");
+    ops.def("dilate_backward(Tensor grad_output, Tensor input, Tensor kernel, int border) -> (Tensor, Tensor)");
 }
 
 // CPU implementation
@@ -20,6 +22,8 @@ TORCH_LIBRARY_IMPL_EXPAND(TORCH_EXTENSION_NAME, CPU, ops) {
 TORCH_LIBRARY_IMPL_EXPAND(TORCH_EXTENSION_NAME, CUDA, ops) {
     ops.impl("erode", &serron::erode);
     ops.impl("dilate", &serron::dilate);
+    ops.impl("erode_backward", &serron::erode_backward);
+    ops.impl("dilate_backward", &serron::dilate_backward);
 }
 #endif